Added an option to trigger spellcast on the down portion of a click
This option can be found in the Clique UI, under the 'options' panel.  When this option is active, any buttons that register for click-casting will be registered for 'AnyDown', which will cause spells to trigger on the down portion of the click.  By default, Clique continues to function as usual, triggering spellcasts on the up portion of a click.