Add an option to trigger 'combat' changed based on party members
This option, when enabled, will do a bit more work trying to check for
when your party or raid members enter combat. When this happen, your
click set will be changed and your 'ooc' clicks will be removed, even
though you aren't strictly in combat. This feature is very experimental,
and should be enabled with care.