Revamp of attribute calculation/setting
The logic seems to be a bit more straightforward and it seems that the
ooc/friend/enemy layering is working properly. The major changes are:

  * Clique will only calculate new attributes when you change bindings
  * Clique will apply the attributes when bindings change
  * If you have ooc bindings, then they will swap on enter/leave combat
  * Otherwise, they will not change, so there's no overhead from that

out-of-combat bindings should now correctly be applied over any other
type of binding with the same key. If this isn't the case, please report
it to me.