Attempt to fix the dangling bindings issue
If the player is over a frame that has key bindings active and that
frame is hidden or obscured then those bindings will remain active until
another Clique-enabled frame is moused over. This fix uses the attribute
driver template to clear this dangling button when the player no longer
has a 'mouseover' unit.

This change could potentially cause problems for small corner cases
where non-unit-frames are enabled for Clique casting, but I do not
currently know of any uses where this is the case.